requirements.txt
inflect==5.0.2
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
six==1.15.0
tqdm==4.56.0
Pillow~=8.1.0
